# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VentasFix API",
        "description": "API REST del backoffice de VentasFix para la gestión de usuarios, productos y clientes. Pensada para el consumo desde aplicaciones de terceros (por ejemplo, el sistema de gestión Softland). Todos los endpoints protegidos requieren un token Bearer obtenido en /api/login.",
        "contact": {
            "name": "Felipe Jiménez",
            "email": "felipeandresjb99@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión y obtener un token",
                "operationId": "1257f11ade1e6ee3a26cfe338badfb8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@ventasfix.cl"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Autenticación exitosa."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef1234567890"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión (revoca el token actual)",
                "operationId": "d0090497298d35d96a9fcde3666ff010",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rut": {
                        "type": "string",
                        "example": "12.345.678-5"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Felipe"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Jiménez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "felipe@ventasfix.cl"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Error de validación",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El campo email es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Introduce el token entregado por /api/login como: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}